Let 404s flow through the error boundary; rename ErrorBoundary.tsx to errors.tsx
Remove the early !match return that bypassed the React render entirely, and make the page script injection conditional so the Router's NotFoundGate mechanism handles unmatched URLs like any other routing error.
